Remove old rendering code and get working on hardware (rendering, at least)

+++ src/utils/file_utils.c
@@ -41,8 +41,8 @@ bool file_copy(const char* to, const char* from) {
     fs_stat(from, &in_stat);
     fs_stat(to, &out_stat);
 
-    printf("File copy: old size: %lld, new size: %lld\n",
-           in_stat.st_size, out_stat.st_size);
+    printf("File copy: old size: %d, new size: %d\n",
+           (unsigned int)in_stat.st_size, (unsigned int)out_stat.st_size);
 
     // Files same size - assume file copy was successful.
     return (in_stat.st_size == out_stat.st_size);
